+++ backend/src/neuroagent/schemas.py
@@ -4,15 +4,6 @@ from typing import ClassVar, Literal
 
 from pydantic import BaseModel, Field
 
-
-class KGMetadata(BaseModel):
-    """Knowledge Graph Metadata."""
-
-    file_extension: str
-    brain_region: str
-    is_lnmc: bool = False
-
-
 Category = Literal[
     "image",
     "json-barplot",
@@ -21,6 +12,29 @@ Category = Literal[
     "json-piechart",
     "json-scatterplot",
     "json-multi-linechart",
+]
+
+EntityRoute = Literal[
+    "analysis-software-source-code",
+    "brain-atlas",
+    "brain-atlas-region",
+    "emodel",
+    "cell-composition",
+    "experimental-bouton-density",
+    "experimental-neuron-density",
+    "experimental-synapses-per-connection",
+    "memodel",
+    "mesh",
+    "me-type-density",
+    "reconstruction-morphology",
+    "electrical-cell-recording",
+    "electrical-recording-stimulus",
+    "single-neuron-simulation",
+    "single-neuron-synaptome",
+    "single-neuron-synaptome-simulation",
+    "ion-channel-model",
+    "subject",
+    "validation-result",
 ]